# Migrate `log_sigmoid_forward` & `log_sigmoid_backward` from TH to Aten (CUDA)