--- a/aten/src/ATen/native/Activation.cpp
+++ b/aten/src/ATen/native/Activation.cpp
@@ -726,7 +726,8 @@ Tensor leaky_relu_backward(
 }
 
 std::tuple<Tensor, Tensor> log_sigmoid_forward(const Tensor& input) {
-  Tensor result;
+  // FIXME: do these actually need to be zeros_like or can they be empty_like?
+  auto result = at::zeros_like(input, at::MemoryFormat::Contiguous);
   auto buffer = at::zeros_like(input, at::MemoryFormat::Contiguous);
   auto iter = TensorIterator::binary_op(result, input, buffer, /*check_mem_overlap=*/true);
   log_sigmoid_stub(iter.device_type(), iter);