--- a/aten/src/ATen/native/cpu/Activation.cpp
+++ b/aten/src/ATen/native/cpu/Activation.cpp
@@ -59,7 +59,7 @@ static void log_sigmoid_kernel(TensorIterator& iter) {
   });
 }
 
-static void log_sigmoid_backward_cpu_kernel(TensorIterator& iter) {
+static void log_sigmoid_backward_kernel(TensorIterator& iter) {
   AT_DISPATCH_FLOATING_TYPES(iter.dtype(), "log_sigmoid_backward_cpu", [&]() {
     using Vec = Vec256<scalar_t>;
     auto zero_val = scalar_t(0);
@@ -595,7 +595,7 @@ void glu_backward_kernel(TensorIterator& iter) {
 } // namespace
 
 REGISTER_DISPATCH(log_sigmoid_stub, &log_sigmoid_kernel);
-REGISTER_DISPATCH(log_sigmoid_backward_cpu_stub, &log_sigmoid_backward_cpu_kernel);
+REGISTER_DISPATCH(log_sigmoid_backward_cpu_stub, &log_sigmoid_backward_kernel);
 REGISTER_DISPATCH(threshold_stub, &threshold_kernel);
 REGISTER_DISPATCH(elu_stub, &elu_kernel);
 REGISTER_DISPATCH(elu_backward_stub, &elu_backward_kernel);