--- a/aten/src/ATen/native/cpu/Activation.cpp
+++ b/aten/src/ATen/native/cpu/Activation.cpp
@@ -52,9 +52,9 @@ inline void _vec_log_sigmoid(Tensor& output, Tensor& buffer, const Tensor& input
 
 static void log_sigmoid_kernel(TensorIterator& iter) {
   AT_DISPATCH_FLOATING_TYPES(iter.dtype(), "log_sigmoid_cpu", [&] {
-    auto output = iter.output();
-    auto input = iter.tensor(1);
-    auto buffer = iter.tensor(2);
+    auto& output = iter.output();
+    auto& input = iter.tensor(1);
+    auto& buffer = iter.tensor(2);
     _vec_log_sigmoid<scalar_t>(output, buffer, input);
   });
 }