--- a/aten/src/ATen/native/native_functions.yaml
+++ b/aten/src/ATen/native/native_functions.yaml
@@ -5658,9 +5658,6 @@
 - func: log_sigmoid_forward(Tensor self) -> (Tensor output, Tensor buffer)
   use_c10_dispatcher: full
   python_module: nn
-  dispatch:
-    CPU: log_sigmoid_forward_cpu
-    CUDA: legacy::cuda::_thnn_log_sigmoid_forward
 
 - func: log_sigmoid_backward.grad_input(Tensor grad_output, Tensor self, Tensor buffer, *, Tensor(a!) grad_input) -> Tensor(a!)
   python_module: nn