… aten/src/ATen/native/cuda/UnaryOpsKernel.cu
… aten/src/ATen/nn.yaml
… aten/src/THCUNN/CMakeLists.txt
… aten/src/THCUNN/LogSigmoid.cu
… aten/src/THCUNN/generic/LogSigmoid.cu
… aten/src/THCUNN/generic/THCUNN.h