Fix #24590, Fix #24591

X-ref to the port of CPU version #30958 by @XiaobingSuper and a later bugfix #36736 by @gchanan 


# Changes

- Use TensorIterator for `log_sigmoid_forward` since the CUDA kernel implementation relies on it. The CPU kernel now retrieves the tensor via `iter.tensor` and CPU implementation remains unchanged. 

- Rename `log_sigmoid_{forward, backward}_cpu_stub` to `log_sigmoid_{forward, backward}_stub`

- Implement `log_sigmoid_kernel` & `log_sigmoid_backward_kernel`

- Use `at::empty_like` instead of `at::zero_like` to initialize `buffer` & `result`.

- Removed `LogSigmoid.cu` from THCUNN.

# Benchmark

- Built with `python setup.py develop` 

- Testing script:

```py
import torch
import torch.nn as nn
import time
import os

torch.manual_seed(0)

def _time():
    return time.time()

device = torch.device(os.getenv('DEVICE'))
m = nn.LogSigmoid()
#warm up
for n in [1, 10, 100, 1000]:
    input = torch.randn(128, n, requires_grad=True, device=device)
    grad_output = torch.randn(128, n, device=device)
    for i in range(1000):
        output = m(input)
        output.backward(grad_output)

for n in [1, 10, 100, 1000]:
    input = torch.randn(128, n, requires_grad=True, device=device)
    grad_output = torch.randn(128, n, device=device)
    fwd_t = 0
    bwd_t = 0
    for i in range(10000):
        t1 = _time()
        output = m(input)
        t2 = _time()
        output.backward(grad_output)
        t3 = _time()
        fwd_t = fwd_t + (t2 - t1)
        bwd_t = bwd_t + (t3 - t2)
    fwd_avg = fwd_t / 10000 * 1000
    bwd_avg = bwd_t / 10000 * 1000
    print("input size(128, %d) forward time is %.2f (ms); backwad avg time is %.2f (ms)."
          % (n, fwd_avg, bwd_avg))

```

- Before (as of [820e81ba09df46ed89f803f587b53a220a8a4757](https://github.com/pytorch/pytorch/tree/820e81ba09df46ed89f803f587b53a220a8a4757)):

```sh
✗ DEVICE=cuda python test.py
input size(128, 1) forward time is 0.08 (ms); backwad avg time is 0.16 (ms).
input size(128, 10) forward time is 0.08 (ms); backwad avg time is 0.16 (ms).
input size(128, 100) forward time is 0.07 (ms); backwad avg time is 0.15 (ms).
input size(128, 1000) forward time is 0.07 (ms); backwad avg time is 0.15 (ms).
```

```sh
✗ DEVICE=cpu python test.py
input size(128, 1) forward time is 0.02 (ms); backwad avg time is 0.03 (ms).
input size(128, 10) forward time is 0.02 (ms); backwad avg time is 0.03 (ms).
input size(128, 100) forward time is 0.04 (ms); backwad avg time is 0.04 (ms).
input size(128, 1000) forward time is 0.09 (ms); backwad avg time is 0.08 (ms).
```

```sh
✗ OMP_NUM_THREADS=1 DEVICE=cpu python test.py
input size(128, 1) forward time is 0.01 (ms); backwad avg time is 0.02 (ms).
input size(128, 10) forward time is 0.02 (ms); backwad avg time is 0.02 (ms).
input size(128, 100) forward time is 0.08 (ms); backwad avg time is 0.02 (ms).
input size(128, 1000) forward time is 0.74 (ms); backwad avg time is 0.17 (ms).
```


- After:

```sh
✗ DEVICE=cuda python test.py
input size(128, 1) forward time is 0.08 (ms); backwad avg time is 0.16 (ms).
input size(128, 10) forward time is 0.08 (ms); backwad avg time is 0.16 (ms).
input size(128, 100) forward time is 0.08 (ms); backwad avg time is 0.16 (ms).
input size(128, 1000) forward time is 0.08 (ms); backwad avg time is 0.16 (ms).
```

```sh
✗ DEVICE=cpu python test.py
input size(128, 1) forward time is 0.02 (ms); backwad avg time is 0.02 (ms).
input size(128, 10) forward time is 0.02 (ms); backwad avg time is 0.03 (ms).
input size(128, 100) forward time is 0.03 (ms); backwad avg time is 0.04 (ms).
input size(128, 1000) forward time is 0.07 (ms); backwad avg time is 0.07 (ms).
```

```sh
✗ OMP_NUM_THREADS=1 DEVICE=cpu python test.py
input size(128, 1) forward time is 0.01 (ms); backwad avg time is 0.02 (ms).
input size(128, 10) forward time is 0.02 (ms); backwad avg time is 0.02 (ms).
input size(128, 100) forward time is 0.08 (ms); backwad avg time is 0.02 (ms).
input size(128, 1000) forward time is 0.71 (ms); backwad avg time is 0.17 (ms).
```

cc: @VitalyFedyunin